export in-memory game state object
