+++ server/src/index.ts
@@ -1,6 +1,5 @@
 import { PORT } from '#config.js'
 import gameRouter from '#routes/game.js'
-import { initGame } from '#utils/state.js'
 import cors from 'cors'
 import dotenv from 'dotenv'
 import express, { Application, Request, Response } from 'express'
@@ -11,9 +10,6 @@ dotenv.config()
 const app: Application = express()
 app.use(express.json())
 app.use(cors())
-
-// Initialize the game state
-initGame()
 
 app.get('/', (req: Request, res: Response) => {
   res.send('Port Royal API')
