+++ server/src/services/gameService.ts
@@ -1,8 +1,7 @@
 import { GameStatus } from '#types.js'
-import { loadState } from '#utils/state.js'
+import { gameStatus } from '#utils/state.js'
 
 const getStatus = (): GameStatus => {
-  const gameStatus: GameStatus = loadState()
   console.log('Game status:', gameStatus)
   return gameStatus
 }
